Move upload file to utils module
